backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
psycopg2-binary
qrcode[pil]
pillow
pytest
pytest-flask
python-dotenv
flask-cors
